Cleaned up the dnd5e client.
