+++ clients/dnd5e/interface.go
@@ -1,8 +1,16 @@
 package dnd5e
 
-import "github.com/fadedpez/dnd5e-api/entities"
+import (
+	"net/http"
+
+	"github.com/fadedpez/dnd5e-api/entities"
+)
 
 type Interface interface {
 	ListRaces() ([]*entities.Race, error)
 	GetRace(key string) (*entities.Race, error)
+}
+
+type httpIface interface {
+	Get(url string) (*http.Response, error)
 }
